numcodecs_linear_quantize

Function reconstruct_into

pub fn reconstruct_into<T: Float + DeserializeOwned, Q: Unsigned>(
    encoded: &[Q],
    decoded: ArrayViewMutD<'_, T>,
    floatify: impl Fn(Q) -> T,
) -> Result<(), LinearQuantizeCodecError>

Reconstruct the linear-quantized encoded array using the floatify closure into the decoded array.

Errors

Errors with

LinearQuantizeCodecError::HeaderDecodeFailed if decoding the header failed

LinearQuantizeCodecError::MismatchedDecodeIntoShape if the decoded array is of the wrong shape