Trait numcodecs_linear_quantize::Float

pub trait Float:
    Copy
    + Serialize
    + DeserializeOwned
    + Sub<Self, Output = Self>
    + Div<Self, Output = Self>
    + Add<Self, Output = Self>
    + Mul<Self, Output = Self>
    + PartialEq {
    const ZERO: Self;
    const ONE: Self;

    // Required methods
    fn minimum(self, other: Self) -> Self;
    fn maximum(self, other: Self) -> Self;
    fn clamp(self, min: Self, max: Self) -> Self;
    fn scale_for_bits(bits: u8) -> Self;
    fn is_finite(self) -> bool;
}

Floating point types.

Required Associated Constants

const ZERO: Self

0.0

const ONE: Self

1.0

Required Methods

fn minimum(self, other: Self) -> Self

Returns the minimum of the two numbers, ignoring NaN.

fn maximum(self, other: Self) -> Self

Returns the maximum of the two numbers, ignoring NaN.

fn clamp(self, min: Self, max: Self) -> Self

Restrict a value to a certain interval unless it is NaN.

fn scale_for_bits(bits: u8) -> Self

Returns exp2(bits) - 1.0

fn is_finite(self) -> bool

Returns true if this number is neither infinite nor NaN.

Object Safety

This trait is not object safe.