numcodecs_linear_quantize

Enum LinearQuantizeBins

#[repr(u8)]
pub enum LinearQuantizeBins {

    _1B1 = 1,
    _1B2 = 2,
    _1B3 = 3,
    _1B4 = 4,
    _1B5 = 5,
    _1B6 = 6,
    _1B7 = 7,
    _1B8 = 8,
    _1B9 = 9,
    _1B10 = 10,
    _1B11 = 11,
    _1B12 = 12,
    _1B13 = 13,
    _1B14 = 14,
    _1B15 = 15,
    _1B16 = 16,
    _1B17 = 17,
    _1B18 = 18,
    _1B19 = 19,
    _1B20 = 20,
    _1B21 = 21,
    _1B22 = 22,
    _1B23 = 23,
    _1B24 = 24,
    _1B25 = 25,
    _1B26 = 26,
    _1B27 = 27,
    _1B28 = 28,
    _1B29 = 29,
    _1B30 = 30,
    _1B31 = 31,
    _1B32 = 32,
    _1B33 = 33,
    _1B34 = 34,
    _1B35 = 35,
    _1B36 = 36,
    _1B37 = 37,
    _1B38 = 38,
    _1B39 = 39,
    _1B40 = 40,
    _1B41 = 41,
    _1B42 = 42,
    _1B43 = 43,
    _1B44 = 44,
    _1B45 = 45,
    _1B46 = 46,
    _1B47 = 47,
    _1B48 = 48,
    _1B49 = 49,
    _1B50 = 50,
    _1B51 = 51,
    _1B52 = 52,
    _1B53 = 53,
    _1B54 = 54,
    _1B55 = 55,
    _1B56 = 56,
    _1B57 = 57,
    _1B58 = 58,
    _1B59 = 59,
    _1B60 = 60,
    _1B61 = 61,
    _1B62 = 62,
    _1B63 = 63,
    _1B64 = 64,

}

Number of bins for quantization, written in base-2 scientific notation.

The binary #[repr(u8)] value of each variant is equivalent to the binary logarithm of the number of bins, i.e. the binary precision or the number of bits used.

Variants

_1B1 = 1

_1B2 = 2

_1B3 = 3

_1B4 = 4

_1B5 = 5

_1B6 = 6

_1B7 = 7

_1B8 = 8

_1B9 = 9

_1B10 = 10

_1B11 = 11

_1B12 = 12

_1B13 = 13

_1B14 = 14

_1B15 = 15

_1B16 = 16

_1B17 = 17

_1B18 = 18

_1B19 = 19

_1B20 = 20

_1B21 = 21

_1B22 = 22

_1B23 = 23

_1B24 = 24

_1B25 = 25

_1B26 = 26

_1B27 = 27

_1B28 = 28

_1B29 = 29

_1B30 = 30

_1B31 = 31

_1B32 = 32

_1B33 = 33

_1B34 = 34

_1B35 = 35

_1B36 = 36

_1B37 = 37

_1B38 = 38

_1B39 = 39

_1B40 = 40

_1B41 = 41

_1B42 = 42

_1B43 = 43

_1B44 = 44

_1B45 = 45

_1B46 = 46

_1B47 = 47

_1B48 = 48

_1B49 = 49

_1B50 = 50

_1B51 = 51

_1B52 = 52

_1B53 = 53

_1B54 = 54

_1B55 = 55

_1B56 = 56

_1B57 = 57

_1B58 = 58

_1B59 = 59

_1B60 = 60

_1B61 = 61

_1B62 = 62

_1B63 = 63

_1B64 = 64